Refactor setupGraph into TestGraph to reduce code duplication
